Delete old relocator

+++ greffe_cli/include/patch/arch/IArchStubs.hpp
@@ -17,7 +17,7 @@ class IArchStubs {
         virtual std::vector<uint8_t> build_shared_stub(uint64_t at) = 0;
         virtual std::vector<uint8_t> trampoline_init(uint64_t at,
                                                      uint64_t shstub_addr,
-                                                     uint32_t **ptr_array) = 0;
+                                                     uint8_t  **ptr_array) = 0;
 
         virtual std::string_view     name()            const = 0;
         virtual uint8_t              instr_alignment() const = 0;
